srcs/back/requirements.txt
fastapi

uvicorn[standard]

pydantic


sqlalchemy
psycopg2-binary
python-dotenv

bcrypt 


pydantic[email]

pytest

httpx


python-jose[cryptography]
passlib[bcrypt]

pyjwt
